Fix view type mismatch (vw_dashboard_kpis total_clientes)

+++ migrations/012_create_views_crm_intelligence_v2.sql
@@ -11,7 +11,7 @@ SELECT
   COALESCE(sum(COALESCE(p.total, 0)), 0)::numeric AS faturamento_total,
   COALESCE(count(p.*), 0)::bigint AS total_pedidos,
   (COALESCE(sum(COALESCE(p.total, 0)), 0) / NULLIF(count(p.*), 0))::numeric AS ticket_medio,
-  (SELECT count(*)::int FROM public.v2_clientes c)::int AS total_clientes,
+  (SELECT count(*)::bigint FROM public.v2_clientes c)::bigint AS total_clientes,
   (SELECT avg(COALESCE(c.ltv, c.total_gasto, 0))::numeric FROM public.v2_clientes c)::numeric AS ltv_medio
 FROM public.v2_pedidos p
 WHERE p.data_pedido IS NOT NULL;
